Learn search page: Replace tab filters with select filters
