+++ kolibri/plugins/learn/assets/src/views/explore-page/index.vue
@@ -13,20 +13,10 @@
       {{ topic.description }}
     </p>
 
-    <ui-select
-      :label="$tr('display')"
-      :options="filterOptions"
-      v-model="selectedFilter"
-      class="filter"
-    />
-
-    <span class="visuallyhidden" v-if="subtopics.length">{{ $tr('navigate') }}</span>
-
     <content-card-grid :contents="contents" v-if="contents.length">
 
       <template scope="content">
         <content-card
-          v-show="selectedFilter.value === 'all' || selectedFilter.value === content.kind"
           :key="content.id"
           :title="content.title"
           :thumbnail="content.thumbnail"
@@ -52,65 +42,23 @@
   import pageHeader from '../page-header';
   import contentCard from '../content-card';
   import contentCardGrid from '../content-card-grid';
-  import uiSelect from 'keen-ui/src/UiSelect';
   export default {
     $trNameSpace: 'learnExplore',
     $trs: {
       explore: 'Topics',
       navigate: 'Navigate content using headings',
-      all: 'All content',
-      topics: 'Topics',
-      exercises: 'Exercises',
-      videos: 'Videos',
-      audio: 'Audio',
-      documents: 'Documents',
-      html5: 'HTML5 Apps',
-      display: 'Display',
     },
     components: {
       pageHeader,
       contentCard,
       contentCardGrid,
-      uiSelect,
     },
-    data: () => ({ selectedFilter: '' }),
     computed: {
-      filterOptions() {
-        const options = [
-          {
-            label: this.$tr('all'),
-            value: 'all',
-          },
-        ];
-        const kindLabelsMap = {
-          [ContentNodeKinds.TOPIC]: this.$tr('topics'),
-          [ContentNodeKinds.EXERCISE]: this.$tr('exercises'),
-          [ContentNodeKinds.VIDEO]: this.$tr('videos'),
-          [ContentNodeKinds.AUDIO]: this.$tr('audio'),
-          [ContentNodeKinds.DOCUMENT]: this.$tr('documents'),
-          [ContentNodeKinds.HTML5]: this.$tr('html5'),
-        };
-        forEach(kindLabelsMap, (value, key) => {
-          if (this.contentsContain(key)) {
-            options.push({
-              label: value,
-              value: key,
-            });
-          }
-        });
-        return options;
-      },
       title() {
         return this.isRoot ? this.$tr('explore') : this.topic.title;
       },
-      subtopics() {
-        return this.contents.filter(content => content.kind === ContentNodeKinds.TOPIC);
-      },
     },
     methods: {
-      contentsContain(kind) {
-        return some(this.contents, content => content.kind === kind);
-      },
       genLink(node) {
         if (node.kind === ContentNodeKinds.TOPIC) {
           return {
@@ -123,9 +71,6 @@
           params: { channel_id: this.channelId, id: node.id },
         };
       },
-    },
-    mounted() {
-      this.selectedFilter = this.filterOptions[0];
     },
     vuex: {
       getters: {
@@ -146,9 +91,5 @@
     margin-top: 1em
     margin-bottom: 1em
     line-height: 1.5em
-
-  .filter
-    width: 200px
-    margin-top: 2em
 
 </style>
